pyproject.toml
[project]
name = "ycode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "EugeneOreshkov", email = "zermolov7@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.123.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[dotenv,email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.7",
]
